[CA-106558] & [CA-94713]: parameterise NFS request timeout value.
Allow the user to specify the maximum time a VBD should wait before reporting
errors in the case of NFS outages. The new parameter is
SR.other-config:nfs-timeout and it is specified as an integer number of
seconds. This parameter will affect the NFS timeo mount option and the tapdisk
request timeout limit.

Imported-by: Chandrika Srinivasan <chandrika.srinivasan@citrix.com>