CA-107099: Time out in some NFS SR operations
Some control-plane operations on the NFS SR may block for up to half an
hour when the NFS server is not reachable (e.g. sr-scan blocks for 30
mins when blocking the connection to the filer with iptables). Instead of
blocking for that long, wait for a time out to expire (by default 10
secs).