CA-110453: raw iSCSI target provided by domU could not be used
On a XenServer with domU1 and domU2: domU1 is an iSCSI target, dom0 us
using that a a raw iSCSI SR. domU2 is accessing that SR and this
operation fails. This patch adds an additional force_tapdisk parameter
to force the use of tapdisk so that the mentioned configuration works.