CP-5330: Ignore innocuous errors from iscsiadm
When asked for open sessions, recent versions of iscsiadm return a code
different than 0 if list is empty.
Our function exn_on_failure raises an exception if return code is not
zero, so we must trap the exception and continue because an empty list
is fine.

The exception could be handled in a better way but our
exception system needs to be made better.

This patch does not affect old iscsiadm behaviour

Signed-off-by: Germano Percossi <germano.percossi@citrix.com>
Reviewed-by: Zheng Li <zheng.li@citrix.com>

(cherry picked from commit 3af8a6e8 for [CP-8634])
Signed-off-by: Germano Percossi <germano.percossi@citrix.com>