CP-6167: Fix an issue reported by pylint
Abstract class VDI referenced undefined property self.path. We
rely on subclasses to always instantiate this property but there
is no guarantee. Better to use Abstract Base Class module when we
move to Python 2.6.

With this change, the unusual instantiation order in ISOVDI class
was exposed (as it caused us some trouble). we resorted the order
to normal and solved the issue.

Signed-off-by: Zheng Li <zheng.li@eu.citrix.com>