CP-6167: Work around a pylint bug
This is a known bug that pylint misunderstands the return types
of 'stdout'/'stderr' of 'subprocess.Popen.communicate' as lists
(instead of strings). This is because these variables were initially
used as lists inside the function and only being converted to strings
at the end.

The python 3.x lib code doesn't cause this confusion for pylint.
Here we introduce this work around for the time being.

Usually the proper way to workaround pylint limitation is to add
pylint ignore tag around a piece of probmatic code, but in this
circumstance, it would demand us to tag all the callers (lots of)
rather than a single callee here.

[1] http://www.logilab.org/ticket/46273

Signed-off-by: Zheng Li <zheng.li@eu.citrix.com>