CP-6336: Removal of dm-multipath-conf
Migration of dm-multipath-conf.patch content from dm-multipath to sm
with basic conf file.
Previously we patched device-mapper-multipath to use a different
configuration file.

Those changes are now integrated into our own configuration file
that is installed using "update-alternatives" with a backup copy
if the original file is a regular file.

Signed-off-by: Germano Percossi <germano.percossi@citrix.com>
Reviewed-by: Vineeth Thampi Raveendran <vineeth.thampi@citrix.com>
[Robert Breker cherry picked and imported for CA-132062 from commit b0892d4]
Signed-off-by: Robert Breker <robert.breker@citrix.com>