CA-121481: Ensure root disk is not blacklisted if booting from SAN
By default, every device is blacklisted, unless explicitly managed
by the Storage Manager.
When the root disk is managed by multipath, this happens at boot
time and the initrd must be able to add the device to multipath.
This patch ensures that, if the root disk is multipathed, it is
never blacklisted.

This patch can only guarantee that multipath is not broken for
the root disk but it is not responsible for making it multipathed
in the first place.

To make it working at installation time, dracut or mkinitrd should
call it explicitly.

Signed-off-by: Germano Percossi <germano.percossi@citrix.com>