CP-6697: normal ismount can not deal with bind mount
This is a known design issue of bind mount on Linux. As we
now use bind mount for FileSR, we need to fix this. Here we
implement a customized version specific to our environment.

PS. the fix for an issue (CA-125122) caused by a previous
version of this patch is now integrated in this version.

Signed-off-by: Zheng Li <zheng.li@citrix.com>
Reviewed-by: Thanos Makatos <thanos.makatos@citrix.com>