CP-7292: Remove E1101 from the exemption list
E1101: Used when a variable is accessed for a nonexistent member.

Some type magic in a small portion of blktap2.py is too advanced
for pylint to comprehend. So we explicitly tell pylint to ignore
that particular warning on that particular section.

By doing so, we can now enable E1101 for all the rest part of our
code to achieve better safety.

Signed-off-by: Zheng Li <zheng.li@eu.citrix.com>