XOP-591: Dump sm_config as part of NFSVDI.generate_config
This information is used by HA to attach heartbeat VDI when XAPI is
not around. It is necessary to preserve the sm_config information
because it now contains the nosubdir flag which is crucial for
deciding which path to mount.

Previously, we deliberately didn't change this to minimize code
churn, because we knew that Cloudstack didn't use XenServer's HA,
and even if they do, they should use the normal NFSSR with subdirectory.

However, the situation changed. Firstly, CloudStack will use XenServer's
HA; Secondly, a user could set heartbeat SR to CloudStack's primary NFSSR
storage by mistake (which has the nosubdir=true) and he/she will then
stuck in a situation can't easily recover. Therefore, we must fix this
problem properly.

Signed-off-by: Zheng Li <zheng.li@citrix.com>