CP-5962: Revert the commit with the temporary fix for [CA-95530]
There is a problem when raw LUNs containing LVM partitions
created by the guest with a much more recent version of LVM
than the one on Dom0 are scanned by XenServer.
When this problem happens, LVM tools get stuck scanning
consuming 100% of the CPU.

This bug has not been observed any more, so removing the temporary
hack introduced in sm.hg(c099f4902946) in clearwater.

Regardless of the bug, this workaround cannot be in place if
the rawhba SR (aka Borehamwood) is installed by default.

Signed-off-by: Germano Percossi <germano.percossi@citrix.com>