CA-142451: Failed to report all LUNs advertised on targets sharing the same IQN
If a multicontroller array advertises its targets using the same IQN
but different IPs and the LUNs in those targets are distinct,
we are able to show only LUNs coming from the first IP address.

The fix reinstates a line that was removed in commit 391bdf055dff
in our old sm.hg repository.
The purporse of the line is to first check that we have
a valid active iscsi connection using the original pair IP,IQN.
Only if it fails (it should not in my opinion) we fall back to
other ip addresses.

Signed-off-by: Germano Percossi <germano.percossi@citrix.com>
Reviewed-by: Chandrika Srinivasan <chandrika.srinivasan@citrix.com>

(cherry picked from commit 66c89ea0f7f70176fbfcb69654f77afbb2c4111d)
Signed-off-by: Mate Lakat <mate.lakat@citrix.com>