Improving readcache check for file based SRs
The check introduced with commit 4be47e0 happens in the
load function.
This means that it has to be done for the load function
of every file based SR.

Putting the check in the FileSR constructor will make
it available for every derived class, without the need
to add the check for every new SR (as it happened to
me while playing with OCFSSR)

The initial version was using "super" instead of an explicit
call to SR.SR.__init__ but the former proved to be unreliable
where there are many circular imports, failing to identify
objects of the same type just because there id was changed
in subsequent import calls.

Signed-off-by: Germano Percossi <germano.percossi@citrix.com>
[initial review]
Reviewed-by: Thanos Makatos <thanos.makatos@citrix.com>