CA-128095: Unhandled exception in LVMoISCSISR.load()
There is a big try/except block in load() that does not handle all possible
exceptions. If an exception is not handled in place, it is logged and the
function tries to continue immediately after the try/except block. Some
errors are recoverable further down, but not all.

When an exception is thrown, log it and save it; if immediatelly after the
try/except block an 'IndexError' exception is raised, catch it and throw
an 'SROSError' one instead.

Signed-off-by: Kostas Ladopoulos <konstantinos.ladopoulos@citrix.com>
Reviewed-by: Germano Percossi <germano.percossi@citrix.com>