CA-92675: Change ISCSI SR scheduler to NOOP
The ISCSI SR works on a LUNperVDI model, attaching a single LUN to a guest
directly via blkback (or tapdisk, in the Rackspace case). Given that a single
task will be issuing requests at any time, it does not make sense to use the
CFQ scheduler for the underlying block devices.

This patch changes the block device (or all block devices, for the multipath
cases, including the device mapper) scheduler(s) to "noop" for the LUNs
related to the SR.

Signed-off-by: Felipe Franciosi <felipe@paradoxo.org>