CA-155766 Fix race between lvchange and tap-ctl open
The new lvm2 code shipped with CentOS 7 has a different
behaviour from the previous one since it de-activates
a logical volume while changing permission from "rw" to
"r". This change implies that if 'lvchange -p r' is executed
whlile a tap-ctl open is happening a volume in the chain can
just disappear and the open fails with the error mentioned
in CA-155766.

I have enforced:

- a lv lock for every lvchange -p
- a lock for every lv in the chain while executing
  tap-ctl open

Signed-off-by: Stefano Panella <stefano.panella@citrix.com>