CA-180792: Enable ALUA autodetection and group_by_prio for VPLEX
In CA-141002 we have enabled ALUA autodetection for VNX. This commit enables
the same capability as well for VPLEX.
In addition to enable the autodetection, this commit also changes the
path_grouping_policy to group_by_prio, as the current upstream setting
multibus wouldn't make sense for ALUA.

The partner has completed the qualification of the updated multipath
configuration with their storage arrays. They have tested both, ALUA and
legacy mode.

Signed-off-by: Robert Breker <robert.breker@citrix.com>