Redesign multipath handling
The old way of managing multipath. by blacklisting/whitelisting
devices on the fly and trigger a reconfigure each time, was
error prone and opened the door to other bugs, in part due
to some problem in multipath reconfigure itself.

Here we exploited the way multipath works when the option
"find_multipath" is set to yes.

When this option is set, multipath does not manage paths
not present in the wwids file.
This file is populated by multipathd when some conditions
are met.
We ship with a patched multipath version that prevents multipathd
from populating this file so, unless this file is populated
by other means, multipath is not able to manage those paths
even if they are not blacklisted.

So, now, instead of touching multipath.conf to control which
paths can or cannot be managed, we use the multipath CLI
to alter the wwids file.
This avoids us the need of invoking reconfigure and have
custom tools to touch multipath files.

The script invoked by the installer is modified accordingly.

This implementation does not work without our custom multipath
version.

Signed-off-by: Germano Percossi <germano.percossi@citrix.com>
Signed-off-by: Stefano Panella <stefano.panella@citrix.com>