CA-174964: VM.start failed, global name 'XenAPI' is not defined
This problem occurs with VMs that went down with physical host suddenly
stopping(crash/power outage etc). If such VM was exported and then this
pool had host removed post exporting a VM, or if VM is imported to another
pool and it still has host_OpaqueRef in the VDI sm_record, this VM will
fail to import

Mark Starikov identified this issue and initially  proposed a solution that
removes host_OpaqueRef entry while exporting, however we tend to try not to
fiddle with the internal data of the SM backends as much as possible. Hence,
as a fix to this issue a logic is added during reset of a VDI which handles
host_OpaqueRef that is not present in the database.

Signed-off-by: Sharath Babu <sharath.babu@citrix.com>